Codegen: group fns by gen type & reduce vis
