+++ keelhaul/src/codegen.rs
@@ -1,24 +1,18 @@
 //! Generate test cases from types implementing [`TestRegister`]
 
-// Codegen API
-pub use self::test_memory::{gen_memtest_module, MemTestStrategy};
-pub use self::test_register::{TestRegister, ValueOnReset};
+// Public API
+pub use self::test_memory::MemTestStrategy;
+
+// Internal codegen API
+pub(crate) use self::test_memory::gen_memtest_module;
+pub(crate) use self::test_register::{RegTestCases, TestRegister, ValueOnReset};
 
 mod test_memory;
 mod test_register;
 
-use std::{
-    any, cmp,
-    collections::{HashMap, HashSet},
-    fmt, iter, str,
-};
+use std::{any, cmp, collections::HashSet, fmt, iter, str};
 
-use self::test_register::RegTestGenerator;
-use crate::{
-    model::{self, ArchPtr, Registers, UniquePath},
-    TestKind,
-};
-use itertools::Itertools;
+use crate::{model, TestKind};
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 use strum::IntoEnumIterator;
@@ -38,8 +32,8 @@ fn gen_mod_desc() -> String {
 /// * `widest` - Size of the widest register on platform, capable of containing the bits of any
 ///   register on platform.
 fn gen_error_struct(widest: u32, error_derive_debug: bool) -> TokenStream {
-    // It costs a lot of code size to `#[derive(Debug)]` so we only do it if required
-    let opt_derive_debug = if error_derive_debug {
+    // It costs a lot of code size to `#[derive(Debug)]` so we do it only if required
+    let derives = if error_derive_debug {
         quote!(#[derive(Debug)])
     } else {
         quote!()
@@ -77,7 +71,7 @@ fn gen_error_struct(widest: u32, error_derive_debug: bool) -> TokenStream {
         .collect();
 
     quote! {
-        #opt_derive_debug
+        #derives
         pub enum Error {
             #error_variant_defs
         }
@@ -114,42 +108,6 @@ fn gen_preamble(widest: u32, error_derive_debug: bool) -> TokenStream {
             pub uid: &'a str,
         }
     }
-}
-
-/// Place test cases in modules.
-fn create_modules(
-    test_fns_and_defs_by_periph: HashMap<String, Vec<(String, String)>>,
-) -> Vec<TokenStream> {
-    test_fns_and_defs_by_periph
-        .into_iter()
-        .map(|(periph_name, test_fns_and_defs)| {
-            // Separate test functions and definitions into two Vecs
-            let (test_fns, test_defs): (Vec<_>, Vec<_>) = test_fns_and_defs.into_iter().unzip();
-
-            // Create array for test definitions
-            let test_defs_combined: TokenStream = test_defs.join(", ").parse().unwrap();
-            let len = test_fns.len();
-            let value: TokenStream = quote!( [#test_defs_combined] );
-            let test_def_arr = quote! {
-                pub static TEST_CASES: [TestCase; #len] = #value;
-            };
-
-            // Create test functions
-            let mod_name: TokenStream = periph_name.to_lowercase().parse().unwrap();
-            let test_fns_catenated: TokenStream = test_fns.join("").parse().unwrap();
-
-            // Create module with both test definitions and test functions
-            quote! {
-                pub mod #mod_name {
-                    use super::*;
-
-                    #test_def_arr
-
-                    #test_fns_catenated
-                }
-            }
-        })
-        .collect_vec()
 }
 
 #[derive(Clone, Debug)]
@@ -334,84 +292,5 @@ pub(crate) fn bit_count_to_rust_uint_type_str(bit_count: u32) -> &'static str {
         32 => "u32",
         64 => "u64",
         _ => panic!("{bit_count} is not a valid bit count"),
-    }
-}
-
-/// Collection of all test cases for this build.
-pub struct TestCases {
-    pub preamble: TokenStream,
-    pub test_cases: Vec<TokenStream>,
-    test_case_array: TokenStream,
-    pub test_case_count: usize,
-}
-
-impl TestCases {
-    /// Generate test cases for each register.
-    ///
-    /// # Panics
-    ///
-    /// - Failed to parse token stream
-    ///
-    /// # Errors
-    ///
-    /// - Failed to generate test case for a register
-    pub fn from_registers<P: ArchPtr + quote::IdentFragment + 'static>(
-        registers: &Registers<P, model::RefSchemaSvdV1_2>,
-        config: &TestConfig,
-    ) -> Self {
-        let widest = registers.iter().map(|reg| reg.size()).max().unwrap();
-        let preamble = gen_preamble(widest, config.derive_debug);
-
-        let mut test_fns_and_defs_by_periph: HashMap<String, Vec<(String, String)>> =
-            HashMap::new();
-        for register in registers.iter() {
-            let test_gen = RegTestGenerator::from_register(register, config);
-            let test_fn = test_gen.gen_test_fn().to_string();
-            let test_def = test_gen.gen_test_def().to_string();
-            test_fns_and_defs_by_periph
-                .entry(register.top_container_name())
-                .or_default()
-                .push((test_fn, test_def));
-        }
-
-        // Duplicate all test definitions into one big list
-        let test_defs = test_fns_and_defs_by_periph
-            .values()
-            .flatten()
-            .map(|(_func, def)| def)
-            .cloned()
-            .collect_vec();
-        let test_case_count = test_defs.len();
-        let test_defs_combined: TokenStream = test_defs.join(",").parse().unwrap();
-        let test_case_array = quote! {
-            pub static TEST_CASES: [TestCase; #test_case_count] = [ #test_defs_combined ];
-        };
-
-        let test_cases = create_modules(test_fns_and_defs_by_periph);
-
-        Self {
-            preamble,
-            test_cases,
-            test_case_array,
-            test_case_count,
-        }
-    }
-
-    pub fn to_tokens(&self) -> TokenStream {
-        let TestCases {
-            preamble,
-            test_cases,
-            test_case_array,
-            ..
-        } = self;
-        quote! {
-            #preamble
-
-            #(
-                #test_cases
-            )*
-
-            #test_case_array
-        }
     }
 }
