+++ keelhaul/src/codegen/test_register.rs
@@ -1,13 +1,16 @@
+use std::collections::HashMap;
+
 use crate::{
     codegen,
-    model::{self, ArchPtr},
+    model::{self, ArchPtr, UniquePath},
     TestConfig, TestKind,
 };
+use itertools::Itertools;
 use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote};
 
 /// Type that a test can be generated for
-pub trait TestRegister<P>: model::UniquePath {
+pub(crate) trait TestRegister<P>: model::UniquePath {
     /// Get the absolute memory address of the register
     fn addr(&self) -> P;
 
@@ -234,6 +237,121 @@ fn gen_read_is_reset_val_test<P: ArchPtr + quote::IdentFragment + 'static>(
                         })
                 }
             }
+        }
+    }
+}
+
+/// Place test cases in modules.
+fn gen_test_mod_wrapper(
+    test_fns_and_defs_by_periph: HashMap<String, Vec<(String, String)>>,
+) -> Vec<TokenStream> {
+    test_fns_and_defs_by_periph
+        .into_iter()
+        .map(|(periph_name, test_fns_and_defs)| {
+            // Separate test functions and definitions into two Vecs
+            let (test_fns, test_defs): (Vec<_>, Vec<_>) = test_fns_and_defs.into_iter().unzip();
+
+            // Create array for test definitions
+            let test_defs_combined: TokenStream = test_defs.join(", ").parse().unwrap();
+            let len = test_fns.len();
+            let value: TokenStream = quote!( [#test_defs_combined] );
+            let test_def_arr = quote! {
+                pub static TEST_CASES: [TestCase; #len] = #value;
+            };
+
+            // Create test functions
+            let mod_name: TokenStream = periph_name.to_lowercase().parse().unwrap();
+            let test_fns_catenated: TokenStream = test_fns.join("").parse().unwrap();
+
+            // Create module with both test definitions and test functions
+            quote! {
+                pub mod #mod_name {
+                    use super::*;
+
+                    #test_def_arr
+
+                    #test_fns_catenated
+                }
+            }
+        })
+        .collect_vec()
+}
+
+/// Collection of all test cases for this build.
+pub struct RegTestCases {
+    preamble: TokenStream,
+    test_cases: Vec<TokenStream>,
+    test_case_array: TokenStream,
+    pub(crate) test_case_count: usize,
+}
+
+impl RegTestCases {
+    /// Generate test cases for each register.
+    ///
+    /// # Panics
+    ///
+    /// - Failed to parse token stream
+    ///
+    /// # Errors
+    ///
+    /// - Failed to generate test case for a register
+    pub fn from_registers<P: ArchPtr + quote::IdentFragment + 'static>(
+        registers: &model::Registers<P, model::RefSchemaSvdV1_2>,
+        config: &TestConfig,
+    ) -> Self {
+        let widest = registers.iter().map(|reg| reg.size()).max().unwrap();
+        let preamble = codegen::gen_preamble(widest, config.derive_debug);
+
+        let mut test_fns_and_defs_by_periph: HashMap<String, Vec<(String, String)>> =
+            HashMap::new();
+        for register in registers.iter() {
+            let test_gen = RegTestGenerator::from_register(register, config);
+            let test_fn = test_gen.gen_test_fn().to_string();
+            let test_def = test_gen.gen_test_def().to_string();
+            test_fns_and_defs_by_periph
+                .entry(register.top_container_name())
+                .or_default()
+                .push((test_fn, test_def));
+        }
+
+        // Duplicate all test definitions into one big list
+        let test_defs = test_fns_and_defs_by_periph
+            .values()
+            .flatten()
+            .map(|(_func, def)| def)
+            .cloned()
+            .collect_vec();
+        let test_case_count = test_defs.len();
+        let test_defs_combined: TokenStream = test_defs.join(",").parse().unwrap();
+        let test_case_array = quote! {
+            pub static TEST_CASES: [TestCase; #test_case_count] = [ #test_defs_combined ];
+        };
+
+        let test_cases = gen_test_mod_wrapper(test_fns_and_defs_by_periph);
+
+        Self {
+            preamble,
+            test_cases,
+            test_case_array,
+            test_case_count,
+        }
+    }
+
+    pub fn to_tokens(&self) -> TokenStream {
+        let RegTestCases {
+            preamble,
+            test_cases,
+            test_case_array,
+            ..
+        } = self;
+        quote! {
+            #preamble
+
+            #(
+                #test_cases
+            )*
+
+            #test_case_array
         }
     }
 }
